Implement Firestore synchronization and migrate IDs to UUID
This commit integrates Firebase Firestore for remote data persistence and real-time synchronization:
- Adds Google Services plugin, JSON configuration, and necessary manifest permissions (`INTERNET`, `ACCESS_NETWORK_STATE`).
- Implements `FirestoreSync` helper to manage CRUD operations and real-time listeners on the "products" collection.
- Updates `ProductRepository` to dual-write to Room and Firestore, and listen for remote changes to update the local database.
- Refactors `Product` entity `id` from `Long` to `String`, utilizing `UUID` for generation to support distributed unique identifiers.
- Migrates `ProductDao`, ViewModels, Navigation arguments, and UI screens (`ProductListScreen`, `EditProductScreen`) to support String-based IDs.
- Updates Room database version to 3 and enables destructive migration.

+++ app/src/main/java/com/ghaith/smartshop/ui/navigation/AppNavHost.kt
@@ -40,7 +40,7 @@ fun AppNavHost(
             route = "edit_product/{id}"
         ) { backStackEntry ->
 
-            val id = backStackEntry.arguments?.getString("id")!!.toLong()
+            val id = backStackEntry.arguments?.getString("id")!!
 
             EditProductScreen(
                 vm = vm,
